* Skills and buff items should now expire properly.  * Refactored StatusEffectTools and the Player class' interface for applying effects. StatusEffectTools now only contains common methods that all types of status effects (i.e. buff skills, buff items, debuffs from monsters). The specific behaviors for the aforementioned status effects are now split into their own classes.  * Mob skill IDs are now represented internally as 2-bytes (short) instead of 4-bytes (int). In order to simplify compiling of Skill.wz in KvjCompiler, mob skill IDs are still 4-bytes on the KVJ files (and are just downcasted).  * Fixed unable to warp bug following a skill cast.  * Fixed reactor and player drops being dropped from the wrong place if a mob with the same entity ID was visible on the map.  * Added more definitions to PlayerJob.  * Made some progress on ranged attacks (still nothing on magic attacks yet).  * Rechargeable ammo (stars, bullets) are now a special case in the inventory manipulation functions in InventoryTools. Adding rechargeable items to the inventory will not change the quantity of non-full items that share the same item ID, and removing rechargeable items will not clear the slot if the new quantity equals 0.  * Quest requirement watching in Player should perform a bit better now.

+++ src/argonms/loading/skill/McdbSkillDataLoader.java
@@ -57,7 +57,7 @@ public class McdbSkillDataLoader extends SkillDataLoader {
 		skillStats.put(Integer.valueOf(skillid), stats);
 	}
 
-	protected void loadMobSkill(int skillid) {
+	protected void loadMobSkill(short skillid) {
 		Connection con = DatabaseConnection.getWzConnection();
 		MobSkillStats stats = null;
 		try {
@@ -73,7 +73,7 @@ public class McdbSkillDataLoader extends SkillDataLoader {
 		} catch (SQLException e) {
 			LOG.log(Level.WARNING, "Could not read MCDB data for mob skill " + skillid, e);
 		}
-		mobSkillStats.put(Integer.valueOf(skillid), stats);
+		mobSkillStats.put(Short.valueOf(skillid), stats);
 	}
 
 	public boolean loadAll() {
@@ -93,7 +93,7 @@ public class McdbSkillDataLoader extends SkillDataLoader {
 			rs = ps.executeQuery();
 			while (rs.next()) {
 				MobSkillStats stats = new MobSkillStats();
-				mobSkillStats.put(doMobWork(rs, stats, con), stats);
+				mobSkillStats.put(Short.valueOf(doMobWork(rs, stats, con)), stats);
 			}
 			return true;
 		} catch (SQLException ex) {
@@ -112,7 +112,7 @@ public class McdbSkillDataLoader extends SkillDataLoader {
 	}
 
 	public boolean canLoadPlayerSkill(int skillid) {
-		if (skillStats.containsKey(skillid))
+		if (skillStats.containsKey(Integer.valueOf(skillid)))
 			return true;
 		Connection con = DatabaseConnection.getWzConnection();
 		boolean exists = false;
@@ -130,8 +130,8 @@ public class McdbSkillDataLoader extends SkillDataLoader {
 		return exists;
 	}
 
-	public boolean canLoadMobSkill(int skillid) {
-		if (mobSkillStats.containsKey(skillid))
+	public boolean canLoadMobSkill(short skillid) {
+		if (mobSkillStats.containsKey(Short.valueOf(skillid)))
 			return true;
 		Connection con = DatabaseConnection.getWzConnection();
 		boolean exists = false;
@@ -201,8 +201,8 @@ public class McdbSkillDataLoader extends SkillDataLoader {
 		return skillid;
 	}
 
-	private int doMobWork(ResultSet rs, MobSkillStats stats, Connection con) throws SQLException {
-		int skillid = rs.getInt(2);
+	private short doMobWork(ResultSet rs, MobSkillStats stats, Connection con) throws SQLException {
+		short skillid = rs.getShort(2);
 		//there's probably another set of buffs and charged for mob skills...
 		do {
 			byte level = rs.getByte(2);
